Fixed a Bug that crashed the app on Dashboard Menu Action
Whenever a session of logged in user was restored. Any action that called the API would force close the application due
to a Null Pointer Exception at the api object. This happend because when the User was being redirected to the DashboardFragment the api object in the MifosApplication Class wasn't being instantiated.

Wrote a fix that now instantiates the api object in the main application class and now if the application is started and the user's login session is restored all the calls work normally.