Fix #190: NullPointerException when coming back from Add savings account page
This was caused by the Fragment being detached from the parent Activity by the time the "success" method was executed. "getActivity()" passed to ArrayAdapter was null and the following ".getSystemService(String)" call on it was causing the NPE to be thrown. As a solution, we keep reference to the parent activity / context in a class field and pass that to the ArrayAdapter.