MIFOSAC-127: Add charges to Savings Account
A user is now able to add charges to a savings account.

However sometimes the fragment for Savings Account Summary is not added
to the back of the stack so that when the charges list fragment loads
into the container and when the back button is pressed again, the app
crashes. Also in some accounts, the charge type do not load.