Fixing crashes for Transactions (MIFOSAC-140)
This fixes Recurring Accounts withdrawal or payment crashes application
(https://github.com/openMF/android-client/issues/147) as well.

A user will be able to make a withdrawal or make a deposit. The code
will replace all string or character values with empty values so that
the int value can be processed.